requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromadb
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
fastapi==0.115.11
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.158.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.85.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
langchain==0.3.21
langchain-core==0.3.45
langchain-community==0.3.20
langchain-google-genai>=0.8.5
langchain-openai==0.3.9
numpy==1.26.4
openai==1.66.5
orjson==3.10.15
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt==2.7.0
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
rank-bm25==0.2.2
